+++ app/src/main/java/com/sweetmay/advancedcryptoindicators2/view/item/CoinItemView.kt
@@ -5,7 +5,7 @@ import com.sweetmay.advancedcryptoindicators2.utils.converter.Converter
 interface CoinItemView: IItemView {
     fun setIcon(imgUrl: String)
     fun setName(name: String)
-    fun setPrice(price: Float)
+    fun setPrice(price: Double)
     fun setFavIcon(boolean: Boolean)
     fun setPriceChange(change: Converter.ConvertedChange)
 }
